Convert more classes to be exportable
