+++ src/dom/p5.File.js
@@ -1,3 +1,63 @@
+/**
+ * @module DOM
+ * @submodule DOM
+ * @for p5.Element
+ */
+
+import { XML } from '../io/p5.XML';
+
+class File {
+  constructor(file, pInst) {
+    this.file = file;
+
+    this._pInst = pInst;
+
+    // Splitting out the file type into two components
+    // This makes determining if image or text etc simpler
+    const typeList = file.type.split('/');
+    this.type = typeList[0];
+    this.subtype = typeList[1];
+    this.name = file.name;
+    this.size = file.size;
+    this.data = undefined;
+  }
+
+
+  static _createLoader(theFile, callback) {
+    const reader = new FileReader();
+    reader.onload = function (e) {
+      const p5file = new File(theFile);
+      if (p5file.file.type === 'application/json') {
+        // Parse JSON and store the result in data
+        p5file.data = JSON.parse(e.target.result);
+      } else if (p5file.file.type === 'text/xml') {
+        // Parse XML, wrap it in p5.XML and store the result in data
+        const parser = new DOMParser();
+        const xml = parser.parseFromString(e.target.result, 'text/xml');
+        p5file.data = new XML(xml.documentElement);
+      } else {
+        p5file.data = e.target.result;
+      }
+      callback(p5file);
+    };
+    return reader;
+  }
+
+  static _load(f, callback) {
+    // Text or data?
+    // This should likely be improved
+    if (/^text\//.test(f.type) || f.type === 'application/json') {
+      File._createLoader(f, callback).readAsText(f);
+    } else if (!/^(video|audio)\//.test(f.type)) {
+      File._createLoader(f, callback).readAsDataURL(f);
+    } else {
+      const file = new File(f);
+      file.data = URL.createObjectURL(f);
+      callback(file);
+    }
+  }
+}
+
 function file(p5, fn){
   /**
    * A class to describe a file.
@@ -87,57 +147,7 @@ function file(p5, fn){
    * </code>
    * </div>
    */
-  p5.File = class File {
-    constructor(file, pInst) {
-      this.file = file;
-
-      this._pInst = pInst;
-
-      // Splitting out the file type into two components
-      // This makes determining if image or text etc simpler
-      const typeList = file.type.split('/');
-      this.type = typeList[0];
-      this.subtype = typeList[1];
-      this.name = file.name;
-      this.size = file.size;
-      this.data = undefined;
-    }
-
-
-    static _createLoader(theFile, callback) {
-      const reader = new FileReader();
-      reader.onload = function (e) {
-        const p5file = new p5.File(theFile);
-        if (p5file.file.type === 'application/json') {
-          // Parse JSON and store the result in data
-          p5file.data = JSON.parse(e.target.result);
-        } else if (p5file.file.type === 'text/xml') {
-          // Parse XML, wrap it in p5.XML and store the result in data
-          const parser = new DOMParser();
-          const xml = parser.parseFromString(e.target.result, 'text/xml');
-          p5file.data = new p5.XML(xml.documentElement);
-        } else {
-          p5file.data = e.target.result;
-        }
-        callback(p5file);
-      };
-      return reader;
-    }
-
-    static _load(f, callback) {
-      // Text or data?
-      // This should likely be improved
-      if (/^text\//.test(f.type) || f.type === 'application/json') {
-        p5.File._createLoader(f, callback).readAsText(f);
-      } else if (!/^(video|audio)\//.test(f.type)) {
-        p5.File._createLoader(f, callback).readAsDataURL(f);
-      } else {
-        const file = new p5.File(f);
-        file.data = URL.createObjectURL(f);
-        callback(file);
-      }
-    }
-  };
+  p5.File = File;
 
   /**
    * Underlying
@@ -371,6 +381,7 @@ function file(p5, fn){
 }
 
 export default file;
+export { File };
 
 if(typeof p5 !== 'undefined'){
   file(p5, p5.prototype);
